+++ server/component/footer/FooterView.php
@@ -56,13 +56,38 @@ class FooterView extends BaseView
         if (count($options) > 1) {
             //show footer only if there are more than 1 language
             $langOptions = new BaseStyleComponent("select", array(
-                //"label" => "CMS Content Language",
-                "css" => "text-dark smallOverwitten",
+                // "css" => "text-dark smallOverwitten",
                 "value" => $_SESSION['user_language'],
                 "name" => "default_language_locale",
                 "items" => $options,
             ));
             $langOptions->output_content();
+        }
+    }
+
+    /**
+     * Render all themes
+     */
+    private function output_footer_themes()
+    {
+        $options = [];
+        array_push($options, array(
+            "value" => "dark",
+            "text" => "Dark"
+        ));
+        array_push($options, array(
+            "value" => "light",
+            "text" => "Light"
+        ));
+            
+        if (count($options) > 1) {
+            //show footer only if there are more than 1 language
+            $themeOptions = new BaseStyleComponent("select", array(
+                //"label" => "CMS Content Language",
+                "name" => "theme_mode",
+                "items" => $options,
+            ));
+            $themeOptions->output_content();
         }
     }