+++ server/component/footer/tpl_footer.php
@@ -7,5 +7,6 @@
     <div class="footer flex-grow-1">
         <?php $this->output_footer_links(); ?>        
     </div>
+    <div id = "defaultTheme"><?php $this->output_footer_themes(); ?></div>
     <div id = "defaultLanguage"><?php $this->output_footer_languages(); ?></div>
 </div>